+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ## 4.9.13 (unreleased)
 
 
-- Nothing changed yet.
+- Allow passing resource creation metadata to sync upload method
 
 
 ## 4.9.12 (2025-10-16)